fix,test: ModPow2InverseUtilのリファクタとコメント追加,テスト
