--- src/matsu/num/number/modint/InverseModPow2.java
@@ -1,60 +1,0 @@
-/*
- * Copyright © 2025 Matsuura Y.
- * 
- * This software is released under the MIT License.
- * http://opensource.org/licenses/mit-license.php
- */
-
-/*
- * 2025.8.6
- */
-package matsu.num.number.modint;
-
-/**
- * 整数の, 2^(bit) に対するモジュロ逆元を計算することに関するプロトタイプ. <br>
- * 2^(32), 2^(64) に対応し, 符号なし整数として扱う.
- * 
- * @author Matsuura Y.
- */
-final class InverseModPow2 {
-
-    private InverseModPow2() {
-        // インスタンス化不可
-        throw new AssertionError();
-    }
-
-    /**
-     * 符号なし int について, 2^(32) を法とした逆元を計算する.
-     * 
-     * <p>
-     * n は奇数でなければならない.
-     * </p>
-     * 
-     * @param n 符号なし整数
-     * @return n^(-1) (mod 2^(32))
-     */
-    static int invModR(int n) {
-        assert (n & 1) == 1;
-
-        int r = 1;
-        int currentRMask = -1;
-        int currentBitSetMask = 1;
-        int out = 0;
-        while (r != 0) {
-            if ((r & 1) == 1) {
-                // outの注目ビットを立てる
-                out |= currentBitSetMask;
-
-                r -= n;
-            }
-
-            r &= currentRMask;
-            r >>>= 1;
-            currentRMask >>>= 1;
-
-            currentBitSetMask <<= 1;
-        }
-
-        return out;
-    }
-}
